refactor: checkCostAndBalance
