+++ src/modules/admin.ts
@@ -26,19 +26,6 @@ export default class Admin extends ModuleBase {
     const requesterAddress = req.body.auth.addr;
     return await this.depositService.handleDeposit(appName, transferKey, transferValue,requesterAddress);
   }
-
-  /**
-   * Check cost of request and check if account can pay. You should use this function before send or handle request.
-   * If you don't set address, it will use default account's address.
-   * @param {string} appName - App name you want to request service to.
-   * @param {string} serviceName - Service name you want to request to.
-   * @param {string} prompt - Data you want to request to service .
-   * @param {string=} userAddress - Address of account you want to check balance. You should set default account if you don't provide address.
-   * @returns Result cost of service. It throws error when user can't pay.
-   */
-  async checkCostAndBalance(appName: string, serviceName: string, prompt: string, userAddress?: string) {
-    return await this.useService.calculateCostAndCheckBalance(appName, serviceName, prompt, userAddress);
-  }
   
   /**
    * Write response. Then change balance of requester and write history of user balance if response status is success.
