+++ src/modules/app.ts
@@ -150,6 +150,37 @@ export default class App extends ModuleBase {
     return await this.sendTransaction(txBody);
   }
 
+    /**
+   * Check cost of request and check if account can pay. You should use this function before send or handle request.
+   * If you don't set address, it will use default account's address.
+   * @param {string} appName - App name you want to request service to.
+   * @param {string} serviceName - Service name you want to request to.
+   * @param {string} prompt - Data you want to request to service .
+   * @param {string=} userAddress - Address of account you want to check balance. You should set default account if you don't provide address.
+   * @returns Result cost of service. It throws error when user can't pay.
+   */
+    async checkCostAndBalance(appName: string, serviceName: string, value: string, requesterAddress?: string) {
+      requesterAddress = requesterAddress ? requesterAddress : this.getDefaultAccount().address;
+      const billingConfig = await this.getBillingConfig(appName);
+      // TODO(woojae): calculate cost more accurately
+      let serviceBillingConfig = billingConfig.service.default;
+      if(billingConfig.service[serviceName]) {
+        serviceBillingConfig = billingConfig.service[serviceName];
+      }
+      const token = value.split(' ').length;
+      let cost = token * serviceBillingConfig.costPerToken;
+      if (serviceBillingConfig.minCost && cost < serviceBillingConfig.minCost) {
+        cost = serviceBillingConfig.minCost;
+      } else if (serviceBillingConfig.maxCost && cost > serviceBillingConfig.maxCost) {
+        cost = serviceBillingConfig.maxCost;
+      }
+      const balance = await this.getCreditBalance(appName, requesterAddress);
+      if (balance < cost) {
+        throw new Error("not enough balance");
+      }
+      return cost;
+    }
+
   async getCreditBalance(appName: string, userAddress: string) {
     const balancePath = Path.app(appName).balanceOfUser(userAddress);
     return await this.ain.db.ref(balancePath).getValue();
